cmd/common/snapshot: Move common code to dedicated package
