+++ cmd/kubectl-gadget/snapshot/process.go
@@ -15,160 +15,23 @@
 package snapshot
 
 import (
-	"fmt"
-	"sort"
-	"strings"
-
 	"github.com/spf13/cobra"
 
+	commonsnapshot "github.com/kinvolk/inspektor-gadget/cmd/common/snapshot"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
-	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/process-collector/types"
 )
-
-type ProcessFlags struct {
-	showThreads bool
-}
-
-type ProcessParser struct {
-	BaseSnapshotParser
-
-	outputConfig *utils.OutputConfig
-	processFlags *ProcessFlags
-}
 
 func newProcessCmd() *cobra.Command {
 	var commonFlags utils.CommonFlags
-	var processFlags ProcessFlags
+	var processFlags commonsnapshot.ProcessFlags
 
-	cmd := &cobra.Command{
-		Use:   "process",
-		Short: "Gather information about running processes",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			processGadget := &SnapshotGadget[types.Event]{
-				name:        "process-collector",
-				commonFlags: &commonFlags,
-				parser:      NewProcessParser(&commonFlags.OutputConfig, &processFlags),
-				params:      nil,
-			}
-
-			return processGadget.Run()
-		},
+	customRun := func(callback func(traceOutputMode string, results []string) error) error {
+		config := NewSnapshotTraceConfig(commonsnapshot.ProcessGadgetName, commonFlags, nil)
+		return utils.RunTraceAndPrintStatusOutput(config, callback)
 	}
 
-	cmd.PersistentFlags().BoolVarP(
-		&processFlags.showThreads,
-		"threads",
-		"t",
-		false,
-		"Show all threads",
-	)
-
+	cmd := commonsnapshot.NewCommonProcessCmd(&processFlags, &commonFlags.OutputConfig, customRun)
 	utils.AddCommonFlags(cmd, &commonFlags)
 
 	return cmd
-}
-
-func NewProcessParser(outputConfig *utils.OutputConfig, processFlags *ProcessFlags) SnapshotParser[types.Event] {
-	availableCols := map[string]struct{}{
-		"node":      {},
-		"namespace": {},
-		"pod":       {},
-		"container": {},
-		"comm":      {},
-		"tgid":      {},
-		"pid":       {},
-	}
-
-	return &ProcessParser{
-		BaseSnapshotParser: BaseSnapshotParser{
-			availableCols: availableCols,
-		},
-		outputConfig: outputConfig,
-		processFlags: processFlags,
-	}
-}
-
-func (p *ProcessParser) GetColumnsHeader() string {
-	requestedCols := p.outputConfig.CustomColumns
-	if len(requestedCols) == 0 {
-		requestedCols = []string{"node", "namespace", "pod", "container", "comm", "pid"}
-		if p.processFlags.showThreads {
-			requestedCols = []string{"node", "namespace", "pod", "container", "comm", "tgid", "pid"}
-		}
-	}
-
-	return p.BuildSnapshotColsHeader(requestedCols)
-}
-
-func (p *ProcessParser) TransformEvent(e *types.Event) string {
-	var sb strings.Builder
-
-	switch p.outputConfig.OutputMode {
-	case utils.OutputModeColumns:
-		if p.processFlags.showThreads {
-			sb.WriteString(fmt.Sprintf("%s\t%s\t%s\t%s\t%s\t%d\t%d",
-				e.Node, e.Namespace, e.Pod, e.Container,
-				e.Command, e.Tgid, e.Pid))
-		} else {
-			sb.WriteString(fmt.Sprintf("%s\t%s\t%s\t%s\t%s\t%d",
-				e.Node, e.Namespace, e.Pod, e.Container,
-				e.Command, e.Pid))
-		}
-	case utils.OutputModeCustomColumns:
-		for _, col := range p.outputConfig.CustomColumns {
-			switch col {
-			case "node":
-				sb.WriteString(fmt.Sprintf("%s", e.Node))
-			case "namespace":
-				sb.WriteString(fmt.Sprintf("%s", e.Namespace))
-			case "pod":
-				sb.WriteString(fmt.Sprintf("%s", e.Pod))
-			case "container":
-				sb.WriteString(fmt.Sprintf("%s", e.Container))
-			case "comm":
-				sb.WriteString(fmt.Sprintf("%s", e.Command))
-			case "tgid":
-				sb.WriteString(fmt.Sprintf("%d", e.Tgid))
-			case "pid":
-				sb.WriteString(fmt.Sprintf("%d", e.Pid))
-			default:
-				continue
-			}
-			sb.WriteRune('\t')
-		}
-	}
-
-	return sb.String()
-}
-
-func (p *ProcessParser) SortEvents(allProcesses *[]types.Event) {
-	if !p.processFlags.showThreads {
-		allProcessesTrimmed := []types.Event{}
-		for _, i := range *allProcesses {
-			if i.Tgid == i.Pid {
-				allProcessesTrimmed = append(allProcessesTrimmed, i)
-			}
-		}
-		*allProcesses = allProcessesTrimmed
-	}
-
-	sort.Slice(*allProcesses, func(i, j int) bool {
-		pi, pj := (*allProcesses)[i], (*allProcesses)[j]
-		switch {
-		case pi.Node != pj.Node:
-			return pi.Node < pj.Node
-		case pi.Namespace != pj.Namespace:
-			return pi.Namespace < pj.Namespace
-		case pi.Pod != pj.Pod:
-			return pi.Pod < pj.Pod
-		case pi.Container != pj.Container:
-			return pi.Container < pj.Container
-		case pi.Command != pj.Command:
-			return pi.Command < pj.Command
-		case pi.Tgid != pj.Tgid:
-			return pi.Tgid < pj.Tgid
-		default:
-			return pi.Pid < pj.Pid
-		}
-	})
 }
